# film-network-ui/src/app/services/models/film-feedback-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface FilmFeedbackResponse {
  createdAt?: string;
  filmId?: number;
  id?: number;
  lastModifiedAt?: string;
  ownFeedback?: boolean;
  rating?: number;
  review?: string;
  userId?: string;
}