# film-network-ui/src/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI Specification - Luca",
    "description": "OpenAPI documentation for Spring Security",
    "termsOfService": "Terms of Service",
    "contact": {
      "name": "Luca",
      "email": "Luca@example.com"
    },
    "license": {
      "name": "License name",
      "url": "https://example-license.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Local Environment"
    },
    {
      "url": "http://luca.com/film-social-network",
      "description": "Production Environment"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/films/{id}": {
      "get": {
        "tags": [
          "Film"
        ],
        "operationId": "getFilmById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilmResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Film"
        ],
        "operationId": "updateFilm",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FilmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilmResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Film"
        ],
        "operationId": "deleteFilm",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/films": {
      "get": {
        "tags": [
          "Film"
        ],
        "operationId": "getAllFilms",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseFilmResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Film"
        ],
        "operationId": "createFilm",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FilmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/films/rent/{filmId}": {
      "post": {
        "tags": [
          "Film"
        ],
        "operationId": "rentFilm",
        "parameters": [
          {
            "name": "filmId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/films/poster/{filmId}": {
      "post": {
        "tags": [
          "Film"
        ],
        "operationId": "uploadPoster",
        "parameters": [
          {
            "name": "filmId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
...
    "/films/rent/return/{filmId}": {
      "patch": {
        "tags": [
          "Film"
        ],
        "operationId": "returnRentedFilm",
        "parameters": [
          {
            "name": "filmId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/films/rent/approve/{filmId}": {
      "patch": {
        "tags": [
          "Film"
        ],
        "operationId": "approveRentedFilm",
        "parameters": [
          {
            "name": "filmId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/films/archived/{filmId}": {
      "patch": {
        "tags": [
          "Film"
        ],
        "operationId": "updateArchivedStatus",
        "parameters": [
          {
            "name": "filmId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/films/user/borrowed": {
      "get": {
        "tags": [
          "Film"
        ],
        "operationId": "getBorrowedFilms",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseRentedFilmResponse"
                }
              }
            }
          }
        }
      }
    },
    "/films/search": {
      "get": {
        "tags": [
          "Film"
        ],
        "operationId": "searchFilms",
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseFilmResponse"
                }
              }
            }
          }
        }
      }
    },
    "/films/owner/returned": {
      "get": {
        "tags": [
          "Film"
        ],
        "operationId": "getReturnedFilmsForOwner",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseRentedFilmResponse"
                }
              }
            }
          }
        }
      }
    },
    "/films/own": {
      "get": {
        "tags": [
          "Film"
        ],
        "operationId": "getAllFilmsYouOwn",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseFilmResponse"
                }
              }
            }
          }
        }
      }
    },
    "/feedbacks/film/{filmId}": {
      "get": {
        "tags": [
          "Feedback"
        ],
        "operationId": "getFeedbackForFilm",
        "parameters": [
          {
            "name": "filmId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseFilmFeedbackResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "FilmRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "director": {
            "type": "string"
          },
          "imdbId": {
            "type": "string"
          },
          "synopsis": {
            "type": "string"
          },
          "filmPoster": {
            "type": "string"
          },
          "year": {
            "type": "string"
          },
          "archive": {
            "type": "boolean"
          },
          "rating": {
            "type": "number",
            "format": "double"
          },
          "genre": {
            "type": "string"
          }
        },
        "required": [
          "director",
          "genre",
          "imdbId",
          "synopsis",
          "title",
          "year"
        ]
      },
      "FilmResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "director": {
            "type": "string"
          },
          "imdbId": {
            "type": "string"
          },
          "synopsis": {
            "type": "string"
          },
          "filmPoster": {
            "type": "string",
            "format": "byte"
          },
          "year": {
            "type": "string"
          },
          "archive": {
            "type": "boolean"
          },
          "rating": {
            "type": "number",
            "format": "double"
          },
          "genre": {
            "type": "string",
            "enum": [
              "ACTION",
              "ADVENTURE",
              "ANIMATION",
              "BIOGRAPHY",
              "COMEDY",
              "CRIME",
              "DOCUMENTARY",
              "DRAMA",
              "FAMILY",
              "FANTASY",
              "HISTORY",
              "HORROR",
              "MUSICAL",
              "MYSTERY",
              "ROMANCE",
              "SCI_FI",
              "SPORT",
              "THRILLER",
              "WAR",
              "WESTERN"
            ]
          },
          "addedBy": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "FilmFeedbackResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "rating": {
            "type": "number",
            "format": "double"
          },
          "review": {
            "type": "string"
          },
          "filmId": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "string"
          },
          "ownFeedback": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PageResponseFilmResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FilmResponse"
            }
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int64"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
      "PageResponseRentedFilmResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RentedFilmResponse"
            }
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int64"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
      "RentedFilmResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "filmTitle": {
            "type": "string"
          },
          "director": {
            "type": "string"
          },
          "imdbId": {
            "type": "string"
          },
          "rentalDate": {
            "type": "string",
            "format": "date-time"
          },
          "returned": {
            "type": "boolean"
          },
          "returnApproved": {
            "type": "boolean"
          },
          "rentalPrice": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "PageResponseFilmFeedbackResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FilmFeedbackResponse"
            }
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int64"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT authentication scheme",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}